+++ back/myfinanciallife/src/main/java/com/example/myfinanciallife/financialrecord/domain/Income.java
@@ -1,5 +1,6 @@
 package com.example.myfinanciallife.financialrecord.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.example.myfinanciallife.user.domain.User;
@@ -13,7 +14,7 @@ public class Income extends FinancialRecord {
 
     protected Income() {}
 
-    public Income(String description, Double amount, LocalDate date, User user, Category category) {
+    public Income(String description, BigDecimal amount, LocalDate date, User user, Category category) {
         super(description, amount, date, user, category);
     }
